Add try-except around hashlib.md5 calls passing usedforsecurity=False.
An exception is currently thrown if passing the usedforsecurity=False
flag when calling hashlib.md5 using a version of python that currently
does not support FIPS.

At the time of this commit, the RHEL version of python supported
this, however, other versions did not. This issue is captured here:
https://bugs.python.org/issue9216

These changes can be dropped once that issue is resolved.